==> notebooks/3.1 shapefile_preprocessing.ipynb <==
import geopandas as gpd
from shapely.geometry import Point
import os

# Create folders of shapefiles

folder_path_SA2 = '../data/raw/SA2_shapefile'

if not os.path.exists(folder_path_SA2):
    os.makedirs(folder_path_SA2)

folder_path_postcode = '../data/raw/postcode_shapefile'

if not os.path.exists(folder_path_postcode):
    os.makedirs(folder_path_postcode)

# Shapefile of SA2

gdf = gpd.read_file('../data/landing/SA2_2021_AUST_SHP_GDA2020/SA2_2021_AUST_GDA2020.shp')

## Pre-Processing

# Only keep Victoria
vic_gdf = gdf[gdf['STE_NAME21'] == 'Victoria']

# Keep only the desired columns
columns_to_keep = ['SA2_CODE21', 'SA2_NAME21', 'AUS_CODE21', 'AREASQKM21', 'geometry']
vic_gdf = vic_gdf[columns_to_keep]

# Drop the last two rows
vic_gdf = vic_gdf.iloc[:-2]

# Change to int type
vic_gdf['SA2_CODE21'] = vic_gdf['SA2_CODE21'].astype(int)

vic_gdf.head()

output_path = "../data/raw/SA2_shapefile/VIC_SA2_2021_AUST_GDA2020.shp"
vic_gdf.to_file(output_path)

## Shapefile of Postcode

gdf = gpd.read_file('../data/landing/POA_2021_AUST_GDA2020_SHP/POA_2021_AUST_GDA2020.shp')
gdf = gdf.iloc[:-1]
gdf['POA_CODE21'] = gdf['POA_CODE21'].astype(int)
# keep postcode between 3000 and 4000 which is victorial
gdf = gdf[(gdf['POA_CODE21'] >= 3000) & (gdf['POA_CODE21'] <= 4000)]

columns_to_keep = ['POA_CODE21', 'geometry']
gdf = gdf[columns_to_keep]
gdf.head()

output_path = "../data/raw/postcode_shapefile/POA_2021_AUST_GDA2020.shp"
gdf.to_file(output_path)
