==> notebooks/3.2 sa2_postcode.ipynb <==
# Combine Postcode and SA2

import pandas as pd

postcode = pd.read_csv('../data/landing/australian_postcodes.csv')
filtered_vic = postcode[postcode['state'] == 'VIC']
filtered_vic.head()

postcode_to_sa2 = filtered_vic[['postcode', 'SA2_MAINCODE_2016']]
postcode_to_sa2.rename(columns={'SA2_MAINCODE_2016': 'SA2'}, inplace=True)
postcode_to_sa2 = postcode_to_sa2[postcode_to_sa2['SA2'].notnull()]
postcode_to_sa2['SA2'] = postcode_to_sa2['SA2'].astype(int)

# Remove duplicates from postcode_to_sa2 to ensure a one-to-one mapping
postcode_to_sa2 = postcode_to_sa2.drop_duplicates(subset='postcode', keep='first')


# Filter the DataFrame to include only rows with postcodes between 3000 and 4000
filtered_sa2_postcode = postcode_to_sa2[
    (postcode_to_sa2['postcode'] >= 3000) & (postcode_to_sa2['postcode'] <= 4000)
    ]

# Reset the index if needed
filtered_sa2_postcode.reset_index(drop=True, inplace=True)
filtered_sa2_postcode


# Save the DataFrame to a CSV file
filtered_sa2_postcode.to_csv("../data/raw/Processed_SA2_Postcode.csv", index=False)

# Merge with Rental Data

df = pd.read_csv('../data/raw/domain_outliers_removed.csv')

# Merge df and postcode_to_sa2
merged_df = pd.merge(df, filtered_sa2_postcode, on='postcode', how='left')

merged_df.to_csv("../data/curated/domain.csv", index=False)

merged_df.head()
