==> notebooks/6.1 Train_distance.ipynb <==
# Calculate Straight-Line Distance for train

import requests
import pandas as pd
import re
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import time

train = pd.read_csv("../data/landing/datasource-VIC_Govt_PTV-VIC_Govt_DELWP_datavic_PTV_METRO_TRAIN_STATION.csv")
train.head()

def distance(lat1, lon1, lat2, lon2):

    """
    Calculate the approximate distance between two sets of latitude and longitude coordinates using a simplified method.

    Parameters:
    - lat1, lon1: Latitude and longitude of the first location.
    - lat2, lon2: Latitude and longitude of the second location.

    Returns:
    - float: The approximate distance between the two locations in kilometers.
    """
    
    lat_diff = (lat2 - lat1) * 111

    # 1 degree of longitude varies, we use an approximation based on the average latitude
    avg_lat = np.radians((lat1 + lat2) / 2.0)  # Convert average latitude to radians
    lon_diff = (lon2 - lon1) * 111 * np.cos(avg_lat)

    # Calculate the distance using the Pythagorean theorem on the lat_diff and lon_diff
    return np.sqrt(lat_diff**2 + lon_diff**2)


def find_nearest_train(house_lat, house_lon, train):

    """
    Find the nearest distance from a given house's latitude and longitude coordinates to nearby trains.

    Parameters:
    - house_lat: Latitude of the house.
    - house_lon: Longitude of the house.
    - train: DataFrame containing train station locations with 'LATITUDE' and 'LONGITUDE' columns.

    Returns:
    - float: The minimum distance in kilometers to the nearest train.
    """

    # Calculate the distance from the property to each school coordinate
    distances = train.apply(
        lambda row: distance(house_lat, house_lon, row['LATITUDE'], row['LONGITUDE']), 
        axis=1
    )

    # Return the closest distance
    return distances.min()

df = pd.read_csv('../data/raw/domain_outliers_removed.csv')

# Create another column which represents the distance to the nearest train station
df['NearestTrainDistance'] = df.apply(
    lambda row: find_nearest_train(row['Latitude'], row['Longitude'], train), 
    axis=1
)

# Output the csv file
df.to_csv("../data/raw/merge_requirement/Direct_train.csv")
