==> notebooks/8.1 hospital_scraping.ipynb <==
# Get hospital coordinates

import requests
import csv

## Search all hospitals in Victoria

def fetch_all_hospitals_in_victoria():
    """
    Fetch information about all hospitals in the state of Victoria, Australia using the Overpass API.

    Returns:
    - data (dict): A dictionary containing information about hospitals in Victoria.

    This function sends a query to the Overpass API to retrieve data about all hospitals located in 
    the state of Victoria, Australia. It uses the "amenity" tag with the value "hospital" to identify 
    hospital entities.
    """

    # Overpass API endpoint
    url = "http://overpass-api.de/api/interpreter"
    
    # Overpass QL query for all hospitals in Victoria
    query = """
    [out:json];
    area["ISO3166-2"="AU-VIC"]->.searchArea;
    (
      node["amenity"="hospital"](area.searchArea);
      way["amenity"="hospital"](area.searchArea);
      relation["amenity"="hospital"](area.searchArea);
    );
    out center;
    """
    
    response = requests.get(url, params={'data': query})
    data = response.json()

    return data


##  Fetch data

hospitals_data = fetch_all_hospitals_in_victoria()

## store data

def save_to_csv(data, filename):

    """
    Save data to a CSV file.

    Parameters:
    - data: A dictionary containing data to be saved.
    - filename: The name of the CSV file to create and save data to.

    This function takes a dictionary 'data' and saves it to a CSV file with the specified 'filename'. 
    The 'data' dictionary is expected to have a structure where 'elements' is a list of dictionaries, 
    and each dictionary may contain keys such as 'type', 'id', 'latitude', 'longitude', and 'name'.
    """
    
    with open(filename, 'w', newline='') as csvfile:
        fieldnames = ['type', 'id', 'latitude', 'longitude', 'name']
        writer = csv.DictWriter(csvfile, fieldnames=fieldnames)

        writer.writeheader()
        for element in data['elements']:
            name = element['tags']['name'] if 'tags' in element and 'name' in element['tags'] else "Unknown"
            if element['type'] == 'node':
                writer.writerow({'type': 'node', 'id': element['id'], 'latitude': element['lat'], 'longitude': element['lon'], 'name': name})
            elif 'center' in element:
                writer.writerow({'type': element['type'], 'id': element['id'], 'latitude': element['center']['lat'], 'longitude': element['center']['lon'], 'name': name})

# Save data to CSV
save_to_csv(hospitals_data, '../data/landing/all_hospitals_in_victoria.csv')
