==> notebooks/10.2 prepare_modeling_data.ipynb <==
# Data Sepapration

Separate the prepared data to training data (2023) and testing data (2024-2026)

## Read 2023 Data

import pandas as pd
import geopandas as gpd

rental_2023 = pd.read_csv('../data/curated/train/rental_price/rental_price_2023.csv')
income_2023 = pd.read_csv('../data/curated/predict/income_data/income_data_2023.csv')
population_2023 = pd.read_csv('../data/curated/predict/population_data/population_data_2023.csv')
crime_2023 = pd.read_csv('../data/curated/train/crime_data/crime_data_2023.csv')
sa2_postcode = pd.read_csv('../data/raw/Processed_SA2_Postcode.csv')

### Merge Training Data into One Dataframe

merged_df = sa2_postcode.merge(rental_2023, on='SA2', how='left')\
    .merge(income_2023, on='SA2', how='left')\
    .merge(population_2023, on='SA2', how='left')\
    .merge(crime_2023, on='SA2', how='left')

merged_df = merged_df.drop_duplicates()
merged_df = merged_df.dropna(subset=['price'])
merged_df = merged_df.dropna(subset=['avg_income'])
merged_df = merged_df.dropna(subset=['avg_population'])

merged_df['avg_crime'].fillna(0, inplace=True)

merged_df = merged_df.drop_duplicates()

merged_df.to_csv('../data/curated/train/train_data_2023.csv', index=False)

merged_df

## Merge Testing Data into One for Each Year

def test_data_merge(year, sa2_postcode):

    """
    Merges income, population, and crime data for a specific year with SA2 postal 
    code information to create test data.

    Parameters:
        year : The specific year for which test data is required.
        sa2_postcode : The DataFrame containing SA2 postal code information.

    Returns:
        DataFrame: A merged DataFrame containing test data for the given year.
    """

    income = pd.read_csv(f'../data/curated/predict/income_data/income_data_{year}.csv')
    population = pd.read_csv(f'../data/curated/predict/population_data/population_data_{year}.csv')
    crime = pd.read_csv(f'../data/curated/predict/crime_data/crime_data_{year}.csv')

    merged_df = sa2_postcode.merge(income, on='SA2', how='left')\
        .merge(population, on='SA2', how='left')\
        .merge(crime, on='SA2', how='left')

    mean_income = merged_df['avg_income'].mean()
    mean_population = merged_df['avg_population'].mean()

    merged_df['avg_income'].fillna(mean_income, inplace=True)
    merged_df['avg_population'].fillna(mean_population, inplace=True)
    merged_df['avg_crime'].fillna(0, inplace=True)
        
    merged_df.to_csv(f'../data/curated/predict/rental_price/test_data_{year}.csv', index=False)

    return

YEARS = [2024, 2025, 2026]

for year in YEARS:
    test_data_merge(year, sa2_postcode)

pd.read_csv('../data/curated/predict/rental_price/test_data_2024.csv')
