==> notebooks/11.1 combine_results.ipynb <==
# Integrated the Predictions

import pandas as pd
import geopandas as gpd
import folium

ml_model = pd.read_csv('../data/curated/ML_growth_rate.csv')
arima_model = pd.read_csv('../data/curated/ARIMA_growth_rate.csv')

## Take the Mean of Two Predictions

# Concatenate the two DataFrames vertically to combine them
combined_df = pd.concat([ml_model, arima_model], ignore_index=True)

# Group the combined DataFrame by 'postcode' and calculate the mean of 'Growth_Rate'
mean_growth_rate = combined_df.groupby('postcode')['Growth_Rate'].mean().reset_index()
mean_growth_rate['Growth_Rate'] = pd.to_numeric(mean_growth_rate['Growth_Rate'], errors='coerce')
mean_growth_rate = mean_growth_rate.sort_values(by='Growth_Rate', ascending=False)
mean_growth_rate.head()

## Show TOP10 Suburbs

Combine the postcodes with same Growth Rate

grouped = mean_growth_rate.groupby('Growth_Rate')['postcode'].apply(list).reset_index()
sorted_grouped = grouped.sort_values(by='Growth_Rate', ascending=False)
sorted_grouped = sorted_grouped.nlargest(11, 'Growth_Rate')[['postcode', 'Growth_Rate']]
sorted_grouped = sorted_grouped.drop(352) # invalid value

sorted_grouped.to_csv('../data/curated/combined_top10.csv', index=False)
sorted_grouped

## Label the TOP10 Suburbs

gdf = gpd.read_file('../data/landing/POA_2021_AUST_GDA2020_SHP/POA_2021_AUST_GDA2020.shp')
gdf = gdf.drop(gdf.index[-1])
gdf['POA_CODE21'] = gdf['POA_CODE21'].astype(int)

postcodes = sorted_grouped['postcode'].tolist()

m = folium.Map(
    location=[-37.8136, 144.9631],
    tiles='OpenStreetMap',
    zoom_start=10
)

# Iterate through the list of postcodes
for postcode_list in postcodes:
    # Combine the areas
    combined_area = gdf[gdf['POA_CODE21'].isin(postcode_list)].unary_union

    # Calculate the center coordinates of the combined area
    center = combined_area.centroid
    center_coords = [center.y, center.x]  # Latitude and longitude

    # Create a marker with a label
    label = ', '.join(map(str, postcode_list))
    marker = folium.Marker(location=center_coords, popup=label)
    marker.add_to(m)

m.save('../plots/top10_map.html')

m
